LPS-X Move bootstrap classes to the new module
