+++ modules/apps/portal-search-elasticsearch-sidecar/portal-search-elasticsearch-sidecar-api/src/main/java/com/liferay/portal/search/elasticsearch/sidecar/ElasticsearchBootstrapProcessCallable.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.elasticsearch7.internal.bootstrap;
+package com.liferay.portal.search.elasticsearch.sidecar;
 
 import com.liferay.petra.process.ProcessCallable;
 import com.liferay.petra.process.ProcessException;
