+++ modules/apps/portal-search-elasticsearch-sidecar/portal-search-elasticsearch-sidecar-api/src/main/resources/com/liferay/portal/search/elasticsearch/sidecar/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
